+++ assets/scss/modules/_header.scss
@@ -119,7 +119,7 @@
     .header__left {
         max-width: 300px;
         display: flex;
-        // align-items: center;
+        align-items: center;
 
         .menu--product-categories {
             display: none;
@@ -134,9 +134,9 @@
     }
 
     .header__top {
-        // padding: 25px 0;
-        padding-top: 25px;
-        padding-bottom: 10px;
+        padding: 25px 0;
+        // padding-top: 25px;
+        // padding-bottom: 10px;
         background-color: $color-1st;
         border-bottom: 1px solid rgba(0, 0, 0, 0.15);
