+++ components/shared/headers/HeaderDefault.vue
@@ -18,11 +18,6 @@
                 </div>
                 <div class="header__center">
                     <search-header />
-                    <div class="product-types-container" v-if="categories !== undefined">
-                        <span class="product-type-item" v-for="category in categories" :key="category.id" @click="filterProduct(category.name)">
-                            {{ category.name }}
-                        </span>
-                    </div>
                 </div>
                 <div class="header__right">
                     <header-actions />
@@ -48,31 +43,8 @@ export default {
         HeaderActions,
         SearchHeader,
     },
-    computed: {
-        ...mapState({
-            categories: (state) => state.shop.categories,
-        }),
-    },
     mounted() {
         window.addEventListener('scroll', stickyHeader);
-    },
-    methods: {
-        async getProductType() {
-            await this.$store.dispatch('shop/getProductType');
-        },
-
-        filterProduct(name) {
-            if (name === 'All products') {
-                name = ''
-            }
-            this.$store.commit('shop/filterProduct', name);
-            if (window.location.pathname !== '/marketplace') {
-                this.$router.push('/marketplace')
-            }
-        },
-    },
-    created() {
-        this.getProductType();
     },
 };
 </script>
@@ -82,12 +54,5 @@ export default {
     .header__top {
         background-color: $color-1st;
     }
-}
-.product-type-item:hover {
-    cursor: pointer;
-    text-decoration: underline;
-}
-.product-types-container {
-    padding-top: 10px;
 }
 </style>
